Refactored signup and onboarding flow
